Fix unsigned comparison
 read(2) call is defined as returning ssize_t, not unsigned int. Also,
 ixp_write accepts long, which is signed.

+++ cmd/ixpc.c
@@ -28,7 +28,7 @@ usage(void) {
 static void
 write_data(IxpCFid *fid, char *name) {
 	void *buf;
-	uint len;
+	long len;
 
 	buf = emalloc(fid->iounit);;
 	do {
